This change addresses Connector Manaager Issue 130 - Wrong error message is display if an error status is returned by the SetConnectorConfig servlet.
I looked into the case where adding a new connector whose
name is a case-only variant of an existing connector shows a
"No connector configuration returned by the connector manager"
message in the Add Connector page.

The SetConnectorConfigHandler catches the ConnectorExistsException,
and returns an appropriate status code.  Also, since an exception
was thrown, the ConfigureResponse (this.configRes) remains null.

By the observed behaviour, I believe the GSA only checks
for a returned statusId of SUCCESS.  If it is not SUCCESS,
the GSA assumes that ConfigureResponse.message contains
the error message to display.  However, for any of the
5 Exceptions caught by SetConnectorConfigHandler, there
will be a null ConfigureResponse and therefore the GSA shows
"No connector configuration returned by the connector manager"

After some experimentation, I observed that if I return
an empty, but non-null, ConfigureResponse, the correct
error message gets displayed.