This change is the first is a sequence of changes to enhance concurrency control in the connector manager. The design of the overal change is described in the document at - https://docs.google.com/a/google.com/Doc?docid=0AazU_blIkNnpZGMzNDNzcnBfM2R3cHFuY3B4&hl=en.
This change includes the initial implementation of the ConnectorCoordiator
which serves as a sigle synchronization point for operations that reconfigure
the connector and operations that manipulate traversals on the connector. In
addition, this change implements SpringInstantiator based on the
ConnectorCoordinator. Mostly this involves reorganizing and moving of code
needed by the ConnectorCoordinator.

Future changes will
- Enhance ThreadPool/Cancelable to enable calls to ConnectorCoordinator
  when a batch's time runs out.
- Modify MockInstaniator to support ConnectorCoordinator.
- Modify TraversalScheduler to delegate batch control to ConnectorCoordinator 
  and remove batch management logic from TraversalScheduler