Bug 6302590: fix prependCmPrefix for non-XML name='' sequences
These changes were made expressly to be as minimally invasive as
possible, while also trying to be as obviously correct as possible.
Thus, simply because it appears that prependCmPrefix is sometimes called
before removeNestedMarkers, support for correctly handling CDATA was
added to the method, even though in practice prependCmPrefix is only
called on the GSA (and thus, it always does processing after
removeNestedMarkers). Similarly, stripCmPrefix was modified even though
it is not called.

There are three system-visible changes:
1) When trying to find a 'name' attribute in prependCmPrefix, we try to
make sure that we are in an HTML tag. This sometimes produces false
negatives. For example, "<input value='1 > 3' name='example'>" would not
get example replaced with CM_example, as would have been done
previously. This is because the > in '1 > 3' confuses our regular
expression into thinking 'name' isn't within a tag. This shortfall is
unfortunate, but is not readily solvable if we also want things like
"<textarea> name='example' </textarea>" to remain unchanged (which is
what this fix is solving).

2) removeNestedMarkers now converts any CDATA sections it finds to PCDATA
(so < becomes &lt;, and the like) instead of simply removing the start
and end CDATA tags. This was required to fix prependCmPrefix.

3) removeNestedMarkers handles pre-existing nested CDATA sections
differently, as is evident in a modified test. Previously a sequence
like "<![CDATA[a<![CDATA[]]>b]]>" would become "ab", however, now we do
what an XML parser would do and get "a&lt;![CDATA[b]]>". Examining r2397
(the revision the test was introduced) makes it seem that the previous
behavior was not part of the bug fix, and was simply a side-effect of
the chosen implementation.

Code review: http://codereview.appspot.com/6453122/