Fix 3.2.x release branch merges and synchronize the code on the branch with the trunk.
The feed backlog trunk revisions r3161, r3215, and r3254 were
partially merged and partially reverted on both the 3.0.x and 3.2.x
release branches. The end result was that only the
traversal.batch.size was changed on the 3.0.x branch, from 500 to
1000, and the lowering of the feed.backlog.interval from 900 to 120
was inadvertently not included on the 3.2.x branch.

The switch to the FileStore on the 3.2.x branch at 3236 was done
manually, and has whitespace differences from the trunk.

This change was semi-automated using diff and patch. No manual edits
were made.

Code review: http://codereview.appspot.com/17540043/