Add a Base16 class to unify hex string encoding. This replaces four implementations in the connector manager, in PdfUtil, ServletUtil, BasicChecksumGenerator, and CompressedFilterInputStreamTest. The connectors for databases, LDAP, and Notes also have implementations that will be replaced.
The implementation in BasicChecksumGenerator was little-endian, and
the Base16 class is strictly big-endian. This is incompatible, and
diffing connectors that use BasicChecksumGenerator class will reindex
all of their content when first upgraded to a connector manager that
uses the new Base16 class. The little-endian implementation seemed
very unusual and unmotivated, the BasicChecksumGenerator class is no
longer used by any Google connectors, and the Guava BaseEncoding class
that might replace Base16 in the future has no support for
little-endian encodings.

Code review: https://codereview.appspot.com/23280043