Increase the default traversal time limit from 30 minutes to 2 hours (b/18248756).
We continue to see legitimate slowness in the connectors, leading to
support cases that end up just increasing the time limit. We do not
see issues where connectors are hung or slow responding but work OK
when a batch is cancelled and retried. I think the limit is still
philosophically a good thing to have, so I'm just increasing the
limits beyond what we have seen causing problems in the field. We have
been generally recommending 90 minutes and not had any problems.

Code review: https://codereview.appspot.com/166120043/